backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
bcrypt==4.0.1
starlette==0.14.2
pydantic==1.8.2
typing-extensions==4.8.0
anyio==3.7.1
pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1
email-validator==2.1.1 